display actually used font in font picker dialog
requested font may not exist in system, so some replacement will be
used instead, but QFontDialog will try to find non-existing font as
it was requested and as result will set "currently used font" to
app/system default font instead. this doesn't match actual font.

so, use QFontInfo to "resolve" requested font to actually used one
and pass it to dialog instead.